Bug Fixed Acceso a Fases 1 y 2
Se ha solucionado un bug por el cual al intentar acceder a las fases 1 y
2 de las campañas, no se podía por fallos en una consulta a la BDD

+++ IPS2024-PL53/src/main/java/backend/data/accionistas/commands/GetNumAccionesByDniAccionista.java
@@ -21,7 +21,7 @@ public class GetNumAccionesByDniAccionista {
 
 	public int execute() {
 		List<Map<String, Object>> acciones = db.executeQueryMap(QUERY, dni);
-		int numAcciones = (int) acciones.get(0).get("NUM_ACCIONES");
+		int numAcciones = ((Long) acciones.get(0).get("NUM_ACCIONES")).intValue();
 		return numAcciones;
 	}
